+++ docs/source/release_notes/release-notes.rst
@@ -109,6 +109,11 @@ v0.61.0
            Previously, we would log a ``SafeArg`` for these batches that had no content.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/2475>`__)
 
+    *    - |fixed|
+         - SweepResults.getCellTsPairsExamined now returns the correct result when sweep is run over multiple batches. 
+           Previously, the result would only count cell-ts pairs examined in the last batch.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/2830>`__)
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 =======